Add "plugins" menu option in full playback view
- refactor plugin querying in a way that now any
   library activity can utilize it
- Make doc more thorough
- Hopefully fix empty plugin list bug
   (flag was added to wrong intent)
